SDI-88: ♻️  Rename logger
