+++ logger.ts
@@ -1,0 +1,8 @@
+import bunyan from 'bunyan'
+import bunyanFormat from 'bunyan-format'
+
+const formatOut = bunyanFormat({ outputMode: 'short', color: true })
+
+const logger = bunyan.createLogger({ name: 'Check My Diary', stream: formatOut, level: 'debug' })
+
+export default logger
