+++ server.ts
@@ -9,7 +9,7 @@ initialiseAppInsights()
 buildAppInsightsClient()
 
 import { app, metricsApp } from './server/index'
-import logger from './log'
+import logger from './logger'
 
 app.listen(app.get('port'), () => {
   logger.info(`Server listening on port ${app.get('port')}`)
