+++ server/data/dataAccess/db.ts
@@ -1,7 +1,7 @@
 import { Pool, QueryConfig, QueryResult, QueryResultRow } from 'pg'
 import fs from 'fs'
 
-import logger from '../../../log'
+import logger from '../../../logger'
 import config from '../../config'
 
 export const pool = new Pool({
