+++ server/data/hmppsAuthClient.ts
@@ -1,4 +1,4 @@
-import logger from '../../log'
+import logger from '../../logger'
 import config from '../config'
 import RestClient from './restClient'
