+++ server/data/redisClient.ts
@@ -1,6 +1,6 @@
 import { createClient } from 'redis'
 
-import logger from '../../log'
+import logger from '../../logger'
 import config from '../config'
 
 export type RedisClient = ReturnType<typeof createClient>
