+++ server/errorHandler.ts
@@ -1,6 +1,6 @@
 import type { Request, Response, NextFunction } from 'express'
 import type { HTTPError } from 'superagent'
-import logger from '../log'
+import logger from '../logger'
 
 export default function createErrorHandler(production: boolean) {
   return (error: HTTPError, req: Request, res: Response, next: NextFunction): void => {
