+++ server/services/notificationService.ts
@@ -7,7 +7,7 @@ import type {
 import axios from 'axios'
 import moment from 'moment'
 import getSanitisedError from '../sanitisedError'
-import logger from '../../log'
+import logger from '../../logger'
 import baseUrl from '../config'
 
 export default class NotificationService {
